fixing addFile while original file size changes
When adding a file which size is being changed during archiving, the archive would become corrupt because the size in the header is different than the size of the actual data written. So an additional check should be done and if the sizes are different, the header should be updated. A quick test would be to create a loop where you write 1 byte to a test.txt file and then trying to add that file to the archive and extracting the archive.